Add parcel built web folder
