+++ web/index.html
@@ -8,7 +8,6 @@
 
         <link href='https://api.mapbox.com/mapbox-assembly/v0.19.0/assembly.min.css' rel='stylesheet'>
         <script async defer src='https://api.mapbox.com/mapbox-assembly/v0.19.0/assembly.js'></script>
-        <script src='https://unpkg.com/vue'></script>
 
         <script async defer src='https://cdnjs.cloudflare.com/ajax/libs/Turf.js/5.1.5/turf.js'></script>
         <script src='https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.22.1/moment.min.js'></script>
@@ -19,7 +18,7 @@
         <script src='https://api.mapbox.com/mapbox-gl-js/plugins/mapbox-gl-geocoder/v2.2.0/mapbox-gl-geocoder.min.js'></script>
         <link rel='stylesheet' href='https://api.mapbox.com/mapbox-gl-js/plugins/mapbox-gl-geocoder/v2.2.0/mapbox-gl-geocoder.css' type='text/css' />
 
-        <script src='./main.js'></script>
+        <script src='./src/main.js'></script>
     </head>
     <body class='h-full'>
         <div id="app" class="h-full relative">
