+++ web/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "hecate-ui",
+  "version": "1.0.0",
+  "description": "Default Hecate Web UI",
+  "repository": "https://github.com/mapbox/Hecate",
+  "author": "Nick Ingalls <ingalls@protonmail.com>",
+  "license": "MIT",
+  "private": false,
+  "dependencies": {
+    "parcel-bundler": "^1.8.1",
+    "vue": "^2.5.16",
+    "vue-hot-reload-api": "^2.3.0"
+  },
+  "devDependencies": {
+    "@vue/component-compiler-utils": "1.3.1",
+    "vue-template-compiler": "^2.5.16"
+  }
+}
