+++ web/src/main.js
@@ -1,0 +1,9 @@
+import Vue from 'vue';
+import App from './App.vue';
+
+window.onload = () => {
+    new Vue({
+        el: '#app',
+        render: h => h(App)
+    });
+}
